backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.17
simpleeval==1.0.0
python-docx==1.1.2
jinja2==3.1.4
matplotlib==3.9.3
